job control: kill/cont process groups

+++ kernel/sys/syscall.c
@@ -780,7 +780,11 @@ static int sys_shm_release(char * path) {
 }
 
 static int sys_kill(pid_t process, uint32_t signal) {
-	return send_signal(process, signal, 0);
+	if (process < -1) {
+		return group_send_signal(-process, signal, 0);
+	} else {
+		return send_signal(process, signal, 0);
+	}
 }
 
 static int sys_gettimeofday(struct timeval * tv, void * tz) {
